# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Encois Gateway API
  version: 1.0.0
paths:
...
  /api/v1/organization:
    get:
      operationId: getOrganization
      summary: Get the organization structure and scoped details available to the caller
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Organization structure, scoped unit details, visible members, and manageable permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
  /api/v1/organization/onboarding:
    patch:
      operationId: updateOrganizationOnboarding
      summary: Update the organization onboarding configuration
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationOnboardingUpdateRequest'
      responses:
        '200':
          description: Onboarding configuration updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationOnboardingResponse'
        '403': { description: onboarding:manage is required }
  /api/v1/organization/onboarding/start:
    post:
      operationId: startOrganizationOnboarding
      summary: Start the organization Coordinator
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Coordinator accepted and onboarding marked ready for dashboard use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationOnboardingResponse'
        '403': { description: onboarding:manage is required }
        '503': { description: Coordinator or persistence is unavailable }
  /api/v1/organization/units:
    get:
      operationId: listOrganizationUnits
      summary: List the organization structure with caller-specific view and management flags
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Organization units with structural metadata; scoped details are returned only for accessible units
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationUnitListResponse'
    post:
      operationId: createOrganizationUnit
      summary: Create a child organization unit inside the caller's management scope
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationUnitCreateRequest'
      responses:
        '201':
          description: Created organization unit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationUnitResponse'
  /api/v1/organization/members:
    get:
      operationId: listOrganizationMembers
      summary: List organization members visible to the caller's management scope
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Visible organization members
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationMemberListResponse'
  /api/v1/organization/permissions:
    get:
      operationId: listOrganizationPermissions
      summary: List direct membership permissions manageable by the caller
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Manageable direct membership permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationPermissionListResponse'
    post:
      operationId: createOrganizationPermission
      summary: Create or update a direct membership permission
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationPermissionCreateRequest'
      responses:
        '201':
          description: Created or updated direct permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationPermissionResponse'
  /api/v1/organization/permissions/{permissionId}:
    patch:
      operationId: updateOrganizationPermission
      summary: Change a direct membership permission's access level
      security:
        - bearerAuth: []
      parameters:
        - name: permissionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationPermissionUpdateRequest'
      responses:
        '200':
          description: Updated direct permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationPermissionResponse'
    delete:
      operationId: deleteOrganizationPermission
      summary: Remove a direct membership permission
      security:
        - bearerAuth: []
      parameters:
        - name: permissionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Permission removed and audit event recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletedResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    serviceTokenAuth:
      type: apiKey
      in: header
      name: X-Encois-Service-Token
...
  schemas:
...
    OrganizationUnitType:
      type: string
      enum: [organization, department, team, project, service, custom]
    AccessLevel:
      type: string
      enum: [viewer, contributor, manager, admin]
...
    OrganizationMembershipStatus:
      type: string
      enum: [invited, active, suspended]
    OrganizationOnboardingStatus:
      type: string
      enum: [pending, initializing, ready, failed]
    CoordinationMode:
      type: string
      enum: [start-coordinator, connect-only]
    OrganizationUnitProjection:
      type: object
      additionalProperties: false
      required: [id, organizationId, parentId, type, slug, name, description, canView, canManage]
      properties:
        id: { type: string, format: uuid }
        organizationId: { type: string, format: uuid }
        parentId: { type: [string, 'null'], format: uuid }
        type: { $ref: '#/components/schemas/OrganizationUnitType' }
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        canView: { type: boolean, description: The caller may select the unit and read its scoped details. }
        canManage: { type: boolean, description: The caller may manage the unit and its direct permissions. }
        manager: { type: string, description: Present only when canView is true. }
        memberCount: { type: integer, minimum: 0, description: Present only when canView is true. }
    OrganizationMemberProjection:
      type: object
      additionalProperties: false
      required: [id, initials, name, role, roleKey, status]
      properties:
        id: { type: string, format: uuid }
        initials: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        role: { type: string }
        roleKey: { type: string }
        homeUnitId: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/OrganizationMembershipStatus' }
    OrganizationPermissionProjection:
      type: object
      additionalProperties: false
      required: [id, memberId, unitId, access, propagateToChildren]
      properties:
        id: { type: string, format: uuid }
        memberId: { type: string, format: uuid }
        unitId: { type: string, format: uuid }
        access: { $ref: '#/components/schemas/AccessLevel' }
        propagateToChildren: { const: true }
    OrganizationProjection:
      type: object
      additionalProperties: false
      required: [organization, units, members, permissions, onboarding]
      properties:
        organization:
          type: object
          additionalProperties: false
          required: [id, slug, name]
          properties:
            id: { type: string, format: uuid }
            slug: { type: string }
            name: { type: string }
        units:
          type: array
          items: { $ref: '#/components/schemas/OrganizationUnitProjection' }
        members:
          type: array
          items: { $ref: '#/components/schemas/OrganizationMemberProjection' }
        permissions:
          type: array
          items: { $ref: '#/components/schemas/OrganizationPermissionProjection' }
        onboarding: { $ref: '#/components/schemas/OrganizationOnboardingProjection' }
    OrganizationOnboardingProjection:
      type: object
      additionalProperties: false
      required: [organizationId, status, coordinatorId, coordinationMode, selectedWorkflows, createdAt, updatedAt]
      properties:
        organizationId: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/OrganizationOnboardingStatus' }
        coordinatorId: { type: string }
        coordinationMode: { $ref: '#/components/schemas/CoordinationMode' }
        selectedWorkflows:
          description: Published Template keys or current approved Blueprint ids. Runtime and Temporal ids are never accepted here.
          type: array
          maxItems: 50
          items: { type: string, maxLength: 128 }
        lastError: { type: string, maxLength: 1000 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    OrganizationOnboardingUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        coordinationMode: { $ref: '#/components/schemas/CoordinationMode' }
        selectedWorkflows:
          description: Published Template keys or current approved Blueprint ids. The Gateway validates each reference in the organization scope.
          type: array
          maxItems: 50
          items: { type: string, maxLength: 128 }
    OrganizationOnboardingResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/OrganizationOnboardingProjection' }
    OrganizationUnitCreateRequest:
      type: object
      additionalProperties: false
      required: [type, name]
      properties:
        parentId: { type: [string, 'null'], format: uuid }
        type: { $ref: '#/components/schemas/OrganizationUnitType' }
        name: { type: string, minLength: 1, maxLength: 120 }
        slug: { type: string, maxLength: 96 }
    OrganizationPermissionCreateRequest:
      type: object
      additionalProperties: false
      required: [memberId, unitId, access]
      properties:
        memberId: { type: string, format: uuid }
        unitId: { type: string, format: uuid }
        access: { $ref: '#/components/schemas/AccessLevel' }
    OrganizationPermissionUpdateRequest:
      type: object
      additionalProperties: false
      required: [access]
      properties:
        access: { $ref: '#/components/schemas/AccessLevel' }
    OrganizationResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/OrganizationProjection' }
    OrganizationUnitResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/OrganizationUnitProjection' }
    OrganizationUnitListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/OrganizationUnitProjection' }
    OrganizationMemberListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/OrganizationMemberProjection' }
    OrganizationPermissionResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/OrganizationPermissionProjection' }
    OrganizationPermissionListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/OrganizationPermissionProjection' }
    DeletedResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [deleted]
          properties:
            deleted: { const: true }
...